Workaround for import error of org.w3c.com
The issue is caused by an export of
org.w3c.dom.html.HTMLDomImplementation from the Xerces bundle that
creates a split package. A split package is illegal starting with Java
9. A bug (https://issues.apache.org/jira/browse/XERCESJ-1689) has been
filed but no progress so far.

I decided to bundle a patched version (just removed the package) of
Xerces until the issue has been resolved. Afterwards, we have to wait
for the new version to be included in Orbit in order to replace it.